Fix reaturants loading
