+++ src/app/(infoscreen)/@events/error.tsx
