+++ src/app/(infoscreen)/@events/loading.tsx
