+++ src/app/(infoscreen)/@events/page.tsx
