+++ src/app/(infoscreen)/layout.tsx
